+++ src/components/Orderer/interfaces.ts
@@ -1,0 +1,17 @@
+interface IOrdererOptionsSortable {
+  key: string;
+}
+
+export interface IOrdererOptions {
+  sortable?: IOrdererOptionsSortable;
+}
+
+export interface ITableSortOptions {
+  key: string;
+  order: string;
+}
+
+export interface ITableSortProps {
+  sort?: ITableSortOptions;
+  onSortChange?: (key: string, order: string) => void;
+}