+++ src/components/Tooltip/index.stories.tsx
@@ -1,9 +1,9 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/react/types-6-0';
-
-import Tooltip from "./";
+import Tooltip from './';
 import { ITooltipProps } from './interfaces';
 import PublicProfileIcon from '../Icons/PublicProfile';
+import CloseIcon from '../Icons/Close';
 
 export default {
   title: 'Tooltip',
@@ -13,18 +13,58 @@ export default {
 
 const Template: Story<ITooltipProps> = (args) => {
   return (
-    <Tooltip {...args}>
-      <PublicProfileIcon stroke={"#00E394"} />
-    </Tooltip>
-  )
+    <div
+      style={{
+        marginTop: '50px',
+      }}
+    >
+      <PublicProfileIcon data-for={args.id} data-tip />
+      <Tooltip {...args} />
+    </div>
+  );
 };
 
 export const Default = Template.bind({});
-Default.args = { 
-  active: true,
-  backgroundColor: '#192C28',
-  color: '#F3F3F3',
-  content: <span>See public profile</span>,
-  arrow: true, 
-  direction: "right"
+Default.args = {
+  content: 'See public profile',
+  id: 'global',
+  direction: 'top',
+};
+
+const renderContent = () => (
+  <div
+    style={{
+      display: 'flex',
+      flexDirection: 'column',
+      position: 'relative',
+      maxWidth: '31.375rem'
+    }}
+  >
+    <button
+      onClick={Tooltip.hide}
+      style={{
+        display: 'flex',
+        cursor: 'pointer',
+        width: '24px',
+        alignSelf: 'flex-end',
+        background: 'none',
+        border: 'none',
+        color: '#fff',
+        position: 'absolute',
+        top: '-3px',
+        right: '-24px',
+      }}
+    >
+      <CloseIcon fill={'#fff'}/>
+    </button>
+    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse posuere nulla eget nisl lobortis aliquam. Morbi vehicula dapibus nisi, eleifend mattis leo convallis vel. Aenean neque erat, viverra sed lobortis eu, pretium et velit. Proin neque urna, mollis in dictum vel, ornare non est. Proin egestas accumsan diam non bibendum.</p>
+  </div>
+);
+
+export const Clickable = Template.bind({});
+Clickable.args = {
+  content: renderContent(),
+  id: 'global',
+  event: 'click',
+  className: 'clickable'
 };